i9100/i777: Fix default paneldata
The default paneldata in the I9300 update12 source is wrong for the devices
we support, and in fact, in the process of merging it, we wound up creating
uninitialized default data.  This resulted in the kernel crashing when
ld9040.get_lcdtype was 0x0

Instead, use u1_panel_pdata from panel-u1.c like the pre-update12 kernel
as our default.

Change-Id: I22cc5b685f1907e79975e1cdd48c5fb61383172c
Signed-off-by: Andrew Dodd <atd7@cornell.edu>

Conflicts:

	arch/arm/mach-exynos/board-lcd-rgb.c